requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
watchdog>=3.0.0
pytest>=7.0.0
pyyaml>=6.0
jupyter>=1.0.0
